Making it clearer where to place header styling
The example does not show where the Style tag and contents should be placed for this to work. It does not seem to work under ```<Pages.Reources>``` but does so within the ```<controls:DataGrid>``` tag. It also does not clarify that the ```<TextBlock>``` is being used for a header title rather than the column headers themselves. Thus, two options are given.